+++ frontend/src/components/BrandLogo/BrandLogo.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import { BrandLogo, BrandLogoBlack } from './styled';
+
+const Logo: React.FC = () => (
+  <BrandLogo>
+    :i<BrandLogoBlack>Tech</BrandLogoBlack>Art
+  </BrandLogo>
+);
+
+export default Logo;
