+++ frontend/src/components/BrandLogo/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BrandLogo';
