+++ frontend/src/components/BrandLogo/styled.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const BrandLogo = styled('div')({
+  width: '139px',
+  display: 'block',
+  height: '38px',
+  fontSize: '32px',
+  lineHeight: '32px',
+  fontWeight: 'bold',
+  margin: '22px 163px 20px 40px',
+  color: '#D43E41',
+});
+export const BrandLogoBlack = styled('span')({
+  color: 'black',
+});
